Refactoring architecture for AddItem page
