+++ app/src/main/java/com/github/brugapp/brug/view_model/AddItemViewModel.kt
@@ -1,0 +1,48 @@
+package com.github.brugapp.brug.view_model
+
+import android.widget.*
+import androidx.lifecycle.ViewModel
+import com.github.brugapp.brug.fake.MockDatabase
+import com.github.brugapp.brug.model.Item
+
+class AddItemViewModel : ViewModel() {
+
+    fun addItem(itemNameView : EditText, description : EditText){
+        val itemId = 1
+        val newItem = Item(itemNameView.text.toString(), itemId, description.text.toString())
+        MockDatabase.currentUser.addItem(newItem)
+    }
+
+    fun verifyForm(nameHelperText : TextView, itemNameView : EditText) : Boolean{
+
+        //checking if the item's name is valid (i.e non-empty)
+        nameHelperText.text = validName(itemNameView)
+
+        if(nameHelperText.text.isNotEmpty()){
+
+            /* Alert window pops up to explain that name is invalid
+            Currently commented to make testing easier
+            AlertDialog.Builder(this)
+                .setTitle("Invalid Name")
+                .setMessage(nameHelperText.text)
+                .setPositiveButton("Continue"){ _,_ ->
+                }
+                .show()*/
+            return false
+        }
+
+        return true
+    }
+
+     private fun validName(itemNameView : EditText): String {
+        val nameText = itemNameView.text.toString()
+        if(nameText.isEmpty()){
+            return "Name must contain at least 1 character"
+        }
+
+        return ""
+    }
+
+
+
+}
