+++ app/src/main/res/layout/activity_add_item.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".AddItemActivity">
+    tools:context=".ui.AddItemActivity">
 
 
     <TextView
